暂时采用一种性能略耗的方案
