+++ EmmyLua-LS/src/main/kotlin/com/tang/vscode/diagnostics/DiagnosticsOptions.kt
@@ -24,4 +24,6 @@ object DiagnosticsOptions {
     var undeclaredVariable = InspectionsLevel.None
 
     var assignValidation = InspectionsLevel.None
+
+    var deprecated = InspectionsLevel.Warning
 }
